Extract getExtensionContextReturningSingleMethod from ParameterizedTestExtensionTests to TestExtensionContext
